pyproject.toml
[project]
name = "algonim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio[ffmpeg]>=2.37.2",
    "pillow>=11.0.0",
    "pyglet>=2.0.20",
    "pygments>=2.18.0",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
